Moved event related functions into a separate file. Moved blender functionality into a separate file
